Improved table of contents.
Does not take up so much horizontal space on wide screens.
Is sticky on wide screens, it shows on the left as you scroll.
Does not show every single header in the TOC.
Has nested nav headers, but they are not shown currently.
The ordering was incorrect before as well.
If there are no headers, the TOC is hidden.